pyproject.toml
[project]
name = "trading-order-entries"
version = "0.1.0"
description = "Order CLI for Alpaca"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alpaca-py>=0.43.2",
    "asyncio>=4.0.0",
    "dbt-core>=1.11.2",
    "duckdb>=1.4.3",
    "polars>=1.36.1",
    "prompt-toolkit>=3.0.52",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "questionary>=2.1.1",
]

[project.scripts]
trading-order-entries = "trading_order_entries.main:cli"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "coverage>=7.13.0",
    "pytest>=9.0.2",
]
